+++ lib-lightset/include/lightsetchain.h
@@ -42,10 +42,10 @@ public:
 	LightSetChain();
 	~LightSetChain() override;
 
-	void Start(uint8_t nPort) override;
-	void Stop(uint8_t nPort) override;
+	void Start(uint32_t nPort) override;
+	void Stop(uint32_t nPort) override;
 
-	void SetData(uint8_t nPort, const uint8_t *, uint16_t) override;
+	void SetData(uint32_t nPort, const uint8_t *, uint32_t) override;
 
 	void Print() override;
